Moved the files to 'upload' package
